+++ framework/bindings/1262/python.cpp
@@ -1622,7 +1622,7 @@ static int brahms_____import_array(void)
 
 COMPONENT_CLASS_CPP::COMPONENT_CLASS_CPP(EventModuleCreateBindings* emc)
 {
-	ComponentVersion v = {-1, -1};
+	ComponentVersion v = {static_cast<UINT16>(-1), static_cast<UINT16>(-1)};
 	versionComponent = v;
 
 	//	lay in path information
